compiles, still wip
