+++ include/core/fitter/Fitter.h
@@ -1,9 +1,6 @@
 #pragma once
 
 #include <fitter/FitterFwd.h>
-#include <fitter/FitResult.h>
-#include <dataset/SimpleDataset.h>
-#include <mini/Minimizer.h>
 
 #include <vector>
 #include <memory>
@@ -16,7 +13,7 @@ namespace fitter {
             /**
              * @brief Perform a fit and return a fit object containing various information. 
              */
-            [[nodiscard]] virtual std::shared_ptr<FitResult> fit() = 0;
+            [[nodiscard]] virtual std::unique_ptr<FitResult> fit() = 0;
 
             /**
              * @brief Perform a fit and return the minimum function value.
@@ -24,26 +21,10 @@ namespace fitter {
             [[nodiscard]] virtual double fit_chi2_only() = 0;
 
             /**
-             * @brief Plot the fitted function and the measured data points.
-             */
-            [[nodiscard]] virtual FitResult::FitInfo plot() = 0;
-
-            /**
-             * @brief Make a residual plot of the fit.
-             */
-            [[nodiscard]] virtual SimpleDataset plot_residuals() = 0;
-
-            [[nodiscard]] virtual std::shared_ptr<FitResult> get_fit() const = 0;
-
-            /**
              * @brief Get the number of degrees of freedom.
              */
             [[nodiscard]] virtual unsigned int dof() const = 0;
-
-            /**
-             * @brief Get the number of degrees of freedom.
-             */
-            [[nodiscard]] unsigned int degrees_of_freedom() const {return dof();}
+            [[nodiscard]] unsigned int degrees_of_freedom() const {return dof();} //< @copydoc dof
 
             /**
              * @brief Get the total number of data points. 
@@ -51,8 +32,13 @@ namespace fitter {
             [[nodiscard]] virtual unsigned int size() const = 0;
 
             /**
-             * @brief Evaluate the chi2 function for the given parameters.
+             * @brief Evaluate the chi2 for the given parameters.
              */
-            [[nodiscard]] virtual double chi2(const std::vector<double>& params) = 0;
+            [[nodiscard]] double chi2(const std::vector<double>& params);
+
+            /**
+             * @brief Get the residuals for the given parameters.
+             */
+            [[nodiscard]] virtual std::vector<double> get_residuals(const std::vector<double>& params) = 0;
     };
 }
