+++ include/core/plots/All.h
@@ -8,7 +8,4 @@
 #include <plots/PlotDistance.h>
 #include <plots/PlotProfiles.h>
 #include <plots/PlotLandscape.h>
-#include <plots/PlotIntensity.h>
-#include <plots/PlotIntensityFit.h>
-#include <plots/PlotIntensityFitResiduals.h>
 #include <plots/PlotOptions.h>
