+++ source/core/fitter/LinearFitter.cpp
@@ -4,168 +4,66 @@ For more information, please refer to the LICENSE file in the project root.
 */
 
 #include <fitter/LinearFitter.h>
-#include <fitter/SimpleLeastSquares.h>
-#include <math/CubicSpline.h>
-#include <hist/intensity_calculator/DistanceHistogram.h>
-#include <hist/Histogram.h>
-#include <utility/Exceptions.h>
-#include <settings/HistogramSettings.h>
-#include <dataset/Dataset2D.h>
-#include <settings/EMSettings.h>
+#include <fitter/FitResult.h>
 
+#include <cmath>
 #include <cassert>
 
 using namespace fitter;
 
-LinearFitter::LinearFitter() = default;
-LinearFitter::LinearFitter(LinearFitter&& other) : fitted(std::move(other.fitted)), data(std::move(other.data)), I0(other.I0), h(std::move(other.h)) {}
-LinearFitter::LinearFitter(const io::ExistingFile& input) {setup(input);}
-LinearFitter::LinearFitter(const io::ExistingFile& input, std::unique_ptr<hist::DistanceHistogram> h) : h(std::move(h)) {setup(input);}
-LinearFitter::LinearFitter(const SimpleDataset& data) : data(data) {}
-LinearFitter::LinearFitter(const SimpleDataset& data, std::unique_ptr<hist::DistanceHistogram> h) : data(data), h(std::move(h)) {}
-LinearFitter::LinearFitter(std::unique_ptr<hist::DistanceHistogram> data, std::unique_ptr<hist::DistanceHistogram> model) : LinearFitter(std::move(data), std::move(model), Limit(settings::axes::qmin, settings::axes::qmax)) {}
-LinearFitter::LinearFitter(std::unique_ptr<hist::DistanceHistogram> data, std::unique_ptr<hist::DistanceHistogram> model, const Limit& limits) : h(std::move(data)) {
-    model_setup(std::move(model), limits);
-}
-LinearFitter::LinearFitter(std::unique_ptr<hist::DistanceHistogram> model) : LinearFitter(std::move(model), Limit(settings::axes::qmin, settings::axes::qmax)) {}
-LinearFitter::LinearFitter(std::unique_ptr<hist::DistanceHistogram> model, const Limit& limits) {
-    model_setup(std::move(model), limits);
+LinearFitter::LinearFitter(const std::vector<double> data, const std::vector<double> model) : data(data), model(model), inv_sigma(data.size(), 1) {
+    assert(data.size() == model.size() && "LinearFitter::LinearFitter: Data and model must have the same size.");
 }
 
-void LinearFitter::model_setup(std::unique_ptr<hist::DistanceHistogram> model, const Limit& limits) {
-    data = model->debye_transform();
-    data.limit_x(limits);
-    data.simulate_errors();
-    if (I0 > 0) {data.normalize(I0);}
-    if (settings::em::simulation::noise) {data.simulate_noise();}
+LinearFitter::LinearFitter(const std::vector<double> data, const std::vector<double> model, const std::vector<double> errors) : LinearFitter(data, model) {
+    assert(data.size() == errors.size() && "LinearFitter::LinearFitter: Data and errors must have the same size.");
+    for (unsigned i = 0; i < errors.size(); ++i) {
+        inv_sigma[i] = 1./errors[i];
+    }
+}
+
+std::vector<double> LinearFitter::fit_params_only() {
+    double S = 0, Sx = 0, Sy = 0, Sxx = 0, Sxy = 0;
+    for (unsigned i = 0; i < data.size(); ++i) {
+        double inv_sig2 = inv_sigma[i]*inv_sigma[i];
+        S += inv_sig2;
+        Sx += data[i]*inv_sig2;
+        Sy += model[i]*inv_sig2;
+        Sxx += std::pow(data[i], 2)*inv_sig2;
+        Sxy += data[i]*model[i]*inv_sig2;
+    }
+
+    double delta = S*Sxx - Sx*Sx;
+    double a = (S*Sxy - Sx*Sy)/delta;
+    double b = (Sxx*Sy - Sx*Sxy)/delta;
+    double a_err = S/delta;
+    double b_err = Sxx/delta;
+    return {a, b, a_err, b_err};
 }
 
 double LinearFitter::fit_chi2_only() {
-    return chi2({});
+    return chi2(fit_params_only());
 }
 
-std::shared_ptr<FitResult> LinearFitter::fit() {
-    std::vector<double> ym = h->debye_transform().get_counts();
-    std::vector<double> Im = splice(ym);
+std::unique_ptr<FitResult> LinearFitter::fit() {
+    auto p = fit_params_only();
 
-    // we want to fit a*Im + b to Io
-    SimpleDataset fit_data(Im, data.y(), data.yerr());
-    if (I0 > 0) {fit_data.normalize(I0);}
-
-    SimpleLeastSquares fitter(fit_data);
-    fitted = fitter.fit();
-    fitted->add_plots(this);
-
-    return fitted;
+    std::unique_ptr<FitResult> f = std::make_unique<FitResult>();
+    f->parameters = {{"a", p[0], std::sqrt(p[2])}, {"b", p[1], std::sqrt(p[3])}};
+    f->dof = dof();
+    f->fval = chi2(p);
+    f->fevals = 1;
+    return f;
 }
 
-void LinearFitter::normalize_intensity(double new_I0) {
-    if (I0 < 0) {data.normalize(new_I0);} // if y0 has not been set yet, we must rescale the data
-    I0 = new_I0;
-}
-
-FitResult::FitInfo LinearFitter::plot() {
-    if (fitted == nullptr) {throw except::bad_order("IntensityFitter::plot: Cannot plot before a fit has been made!");}
-
-    double a = fitted->get_parameter("a").value;
-    double b = fitted->get_parameter("b").value;
-
-    std::vector<double> ym = h->debye_transform().get_counts();
-    std::vector<double> Im = splice(ym);
-
-    // if we have a I0, we need to rescale the data
-    // double factor = I0/ym[0];
-    // std::transform(Im.begin(), Im.end(), Im.begin(), [&factor] (double y) {return factor*y;});
-
-    // calculate the scaled I model values
-    std::vector<double> I_scaled(data.size()); // spliced data
-    std::vector<double> ym_scaled(ym.size()); // original scaled data
-    std::transform(Im.begin(), Im.end(), I_scaled.begin(), [&a, &b] (double I) {return I*a+b;});
-    std::transform(ym.begin(), ym.end(), ym_scaled.begin(), [&a, &b] (double I) {return I*a+b;});
-
-    // prepare the TGraphs
-    FitResult::FitInfo graphs;
-    graphs.fitted_intensity_interpolated = SimpleDataset(data.x(), I_scaled);
-    graphs.fitted_intensity = SimpleDataset(h->get_q_axis(), ym_scaled);
-    graphs.dataset = SimpleDataset(data.x(), data.y(), data.yerr());
-
-    auto lim = graphs.dataset.get_xlimits();
-    lim.expand(0.05);
-    graphs.fitted_intensity.limit_x(lim);
-    return graphs;
-}
-
-SimpleDataset LinearFitter::plot_residuals() {
-    if (fitted == nullptr) {throw except::bad_order("IntensityFitter::plot_residuals: Cannot plot before a fit has been made!");}
- 
-    double a = fitted->get_parameter("a").value;
-    double b = fitted->get_parameter("b").value;
-
-    std::vector<double> ym = h->debye_transform().get_counts();
-    std::vector<double> Im = splice(ym);
-
-    // calculate the residuals
+std::vector<double> LinearFitter::get_residuals(const std::vector<double>& p) {
     std::vector<double> residuals(data.size());
-    for (int i = 0; i < static_cast<int>(data.size()); ++i) {
-        residuals[i] = ((data.y(i) - a*Im[i]-b)/data.yerr(i));
-    }
-
-    // prepare the dataset
-    std::vector<double> xerr(data.size(), 0);
-    return Dataset2D(data.x(), std::move(residuals), std::move(xerr), data.yerr());
-}
-
-void LinearFitter::set_scattering_hist(std::unique_ptr<hist::DistanceHistogram> h) {
-    assert(h != nullptr && "LinearFitter::set_scattering_hist: Attempting to set a nullptr. This is probably not intended.");
-    this->h = std::move(h);
-}
-
-observer_ptr<hist::DistanceHistogram> LinearFitter::get_scattering_hist() {
-    assert(h != nullptr && "LinearFitter::get_scattering_hist: The scattering histogram has not been set.");
-    return h.get();
-}
-
-double LinearFitter::chi2(const std::vector<double>&) {
-    std::vector<double> ym = h->debye_transform().get_counts();
-    std::vector<double> Im = splice(ym);
-
-    // we want to fit a*Im + b to Io
-    SimpleDataset fit_data(Im, data.y(), data.yerr());
-    if (I0 > 0) {fit_data.normalize(I0);}
-
-    SimpleLeastSquares fitter(fit_data);
-    return fitter.fit_chi2_only();
-}
-
-void LinearFitter::setup(const io::ExistingFile& file) {
-    data = SimpleDataset(file); // read observed values from input file
-}
-
-std::vector<double> LinearFitter::splice(const std::vector<double>& ym) const {
-    std::vector<double> Im(data.size()); // spliced model values
-    math::CubicSpline s(h->get_q_axis(), ym);
     for (unsigned int i = 0; i < data.size(); ++i) {
-        Im[i] = s.spline(data.x(i));
+        residuals[i] = (model[i] - (p[0]*data[i] + p[1]))*inv_sigma[i];
     }
-    return Im;
+    return residuals;
 }
 
-unsigned int LinearFitter::size() const {
-    return data.size();
-}
+unsigned int LinearFitter::dof() const {return data.size() - 2;}
 
-unsigned int LinearFitter::dof() const {
-    return data.size() - 3;
-}
-
-std::shared_ptr<FitResult> LinearFitter::get_fit() const {
-    if (fitted == nullptr) {throw except::bad_order("LinearFitter::get_fit: Cannot get the fit results before a fit has been made!");}
-    return fitted;
-}
-
-void LinearFitter::operator=(LinearFitter&& other) {
-    fitted = std::move(other.fitted);
-    data = std::move(other.data);
-    I0 = other.I0;
-    h = std::move(other.h);
-}
+unsigned int LinearFitter::size() const {return data.size();}
