+++ source/core/plots/CMakeLists.txt
@@ -3,9 +3,6 @@ target_sources(ausaxs_core PRIVATE
 	"PlotDataset.cpp"
 	"PlotDistance.cpp"
 	"PlotHistogram.cpp"
-	"PlotIntensity.cpp"
-	"PlotIntensityFit.cpp"
-	"PlotIntensityFitResiduals.cpp"
 	"PlotLandscape.cpp"
 	"PlotOptions.cpp"
 	"PlotProfiles.cpp"
